# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "tiktok clone",
    "description": "\nA simple tiktok clone API\n\n# Sokket.io is used for real-time features like chats, notifications\n\n## chats real-time: ws://localhost:3000/chat?token=<your_token>\n\n### Client -> Server:\n- **Event name**: `SendMessage`\n- **Data**: `{ recipientId: string, message: string }`\n\n### Server -> Client:\n- **Event name**: `NewMessage`\n- **Data**: `{ senderId: string, recipientId: string, message: string }`\n\n\n## Notifications real-time: ws://localhost:3000/notifications?token=<your_token>\n### Server -> Client:\n- **Event name**: `NewNotification`\n- **Data**: `{ userId: string, message: string }`\n\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://course-swap.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/chats/message/{messageId}": {
      "delete": {
        "tags": [
          "ChatsController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {},
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          }
        }
      }
    },
    "/api/chats/": {
      "get": {
        "tags": [
          "ChatsController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "results": {
                          "default": [],
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "content": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "sender": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "username": {
                                    "type": "string"
                                  },
                                  "avatarUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "username",
                                  "avatarUrl"
                                ],
                                "additionalProperties": false
                              },
                              "receiver": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "username": {
                                    "type": "string"
                                  },
                                  "avatarUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "username",
                                  "avatarUrl"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "id",
                              "content",
                              "createdAt",
                              "sender",
                              "receiver"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "total",
                        "results"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "UnauthorizedErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/chats/{userId}": {
      "get": {
        "tags": [
          "ChatsController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 10,
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "UnauthorizedErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "ChatsController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {},
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "UnauthorizedErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  }
}